+++ test/cloudevents-fast-schema-compilation.test.js
@@ -22,6 +22,7 @@ const Fastify = require('fastify')
 
 // use 'ajv' (dependency of fast-json-stringify') in all tests here
 const Ajv = require('ajv')
+const addFormats = require('ajv-formats') // already installed when installing Ajv itself
 
 // import some common test data
 const {
@@ -57,7 +58,7 @@ test('ensure normal instancing of fast validation (like the one exposed by the p
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err, address) => {
+  fastify.listen({ port: 0 }, (err, address) => {
     t.error(err)
 
     const CloudEvent = fastify.CloudEvent
@@ -75,6 +76,7 @@ test('ensure normal instancing of fast validation (like the one exposed by the p
     const ajv = new Ajv({ coerceTypes: true, removeAdditional: true })
     assert(ajv !== null)
     t.ok(ajv)
+    addFormats(ajv) // enhance ajv validation on some formats
     const ceValidate = ajv.compile(ceSchema)
     assert(ceValidate !== null)
     t.ok(ceValidate)
@@ -187,7 +189,7 @@ test('ensure CloudEvent schema and schema compiler (both exposed by the plugin) 
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err, address) => {
+  fastify.listen({ port: 0 }, (err, address) => {
     t.error(err)
 
     const CloudEvent = fastify.CloudEvent