+++ test/cloudevents.test.js
@@ -27,7 +27,7 @@ test('ensure decorator functions (exposed by the plugin) exists', (t) => {
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen({ port: 0 }, (err) => {
     t.error(err)
 
     // ensure CloudEvent constructor function exist in Fastify decorators ...
@@ -57,7 +57,7 @@ test('ensure isValid and validate works good on undefined and null objects', (t)
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen({ port: 0 }, (err) => {
     t.error(err)
     const CloudEvent = fastify.CloudEvent
     t.ok(CloudEvent)
@@ -90,7 +90,7 @@ test('create some CloudEvent instances (empty, without minimal arguments set or 
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen({ port: 0 }, (err) => {
     t.error(err)
     const CloudEvent = fastify.CloudEvent
     t.ok(CloudEvent)
@@ -133,7 +133,7 @@ test('create some CloudEvent instances (with minimal fields set) and ensure they
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen({ port: 0 }, (err) => {
     t.error(err)
     const CloudEvent = fastify.CloudEvent
     t.ok(CloudEvent)
@@ -227,7 +227,7 @@ test('create two CloudEvent instances with all arguments (mandatory and optional
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen({ port: 0 }, (err) => {
     t.error(err)
     const CloudEvent = fastify.CloudEvent
     t.ok(CloudEvent)
@@ -278,7 +278,7 @@ test('create CloudEvent instances with different kind of data attribute, and ens
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen({ port: 0 }, (err) => {
     t.error(err)
     const CloudEvent = fastify.CloudEvent
     t.ok(CloudEvent)
@@ -412,7 +412,7 @@ test('create CloudEvent instances with data encoded in base64, and ensure the va
   t.teardown(() => { fastify.close() })
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen({ port: 0 }, (err) => {
     t.error(err)
     const CloudEvent = fastify.CloudEvent
     t.ok(CloudEvent)